edited modules, added new helper method module
