+++ lib/stat_tracker.rb
@@ -6,6 +6,7 @@ require_relative './game_team'
 require_relative '../modules/game_statables'
 require_relative '../modules/league_statables'
 require_relative '../modules/team_statables'
+require_relative '../modules/team_stat_helper'
 require_relative '../modules/season_statables'
 
 class StatTracker
@@ -13,6 +14,7 @@ class StatTracker
   include LeagueStatables
   include TeamStatables
   include SeasonStatables
+  include TeamStatHelper
   attr_reader :games, :teams, :game_teams
 
   def initialize(games = {}, teams = {}, game_teams = {})
